Доработал взаимодействие Coroutine flow c ViewModel LiveData
