+++ app/src/main/java/ru/geekbrains/android2/crosszerogame/view/list/OpponentHolder.kt
@@ -2,7 +2,7 @@ package ru.geekbrains.android2.crosszerogame.view.list
 
 import androidx.recyclerview.widget.RecyclerView
 import ru.geekbrains.android2.crosszerogame.App
-import ru.geekbrains.android2.crosszerogame.data.Gamer
+import ru.geekbrains.android2.crosszerogame.xdata.Gamer
 import ru.geekbrains.android2.crosszerogame.databinding.ItemGameBinding
 import ru.geekbrains.android2.crosszerogame.utils.strings.GameStrings
