+++ app/src/main/java/ru/geekbrains/android2/crosszerogame/xdata/GameRepository.kt
@@ -1,4 +1,4 @@
-package ru.geekbrains.android2.crosszerogame.data
+package ru.geekbrains.android2.crosszerogame.xdata
 
 interface GameRepository {
     suspend fun gamer(
