+++ app/src/main/res/values/strings.xml
@@ -2,8 +2,8 @@
     <string name="app_name">Крестики-Нолики</string>
     <string name="back4app_server_url">https://parseapi.back4app.com/</string>
     <!-- Paste BOTH strings as required -->
-    <string name="back4app_app_id">IDx83YYoFGvStpGouDuqZc4g8qF4vKUGuuNkGPcC</string>
-    <string name="back4app_client_key">IrxbE17QQH3RoLGlbdgRfbWewcfTGMxSO26x7c5F</string>
+    <string name="back4app_app_id">EAWN0NncaWQOTxYyzwOurVJIC93MqVC5guOlDDpZ</string>
+    <string name="back4app_client_key">XwhpH0NaenpDqF7vFjilu4KCBUznOKAwggwb52Xh</string>
     <string name="begin_as">Начинать</string>
     <string name="first">первым</string>
     <string name="second">вторым</string>
@@ -25,7 +25,7 @@
     <string name="unavailable_nick">Недоступный ник</string>
     <string name="wait_cross">Ожидается игрок за Крестики</string>
     <string name="wait_zero">Ожидается игрок за Нолики</string>
-    <string name="opponent_info">Противник %s за %s размер поля %s уровень %s</string>
-    <string name="cross">крестики</string>
-    <string name="zero">нолики</string>
+    <string name="opponent_info">%s %s %sx%s уровень %s</string>
+    <string name="cross">+</string>
+    <string name="zero">o</string>
 </resources>
